+++ tests/test_parser.py
@@ -1,5 +1,6 @@
 import os
 import pytest
+from ldfparser.diagnostics import LIN_MASTER_REQUEST_FRAME_ID, LIN_SLAVE_RESPONSE_FRAME_ID
 
 from ldfparser.parser import parse_ldf
 from ldfparser.frame import LinFrame
@@ -36,6 +37,12 @@ def test_load_valid_lin20():
     assert ldf.signal('InternalLightsRequest') is not None
     assert ldf.frame('VL1_CEM_Frm1') is not None
     assert ldf.slave('LSM') is not None
+
+    with pytest.raises(LookupError):
+        ldf.get_unconditional_frame('VL1_CEM_Frm1234')
+
+    with pytest.raises(TypeError):
+        ldf.get_unconditional_frame(['VL1_CEM_Frm1'])
 
 
 @pytest.mark.unit
@@ -156,3 +163,20 @@ def test_load_valid_lin_encoders():
     assert converter.name == 'AsciiEncoding'
     assert len(converter._converters) == 1
     assert isinstance(converter._converters[0], ASCIIValue)
+
+@pytest.mark.unit
+def test_load_valid_diagnostics():
+    path = os.path.join(os.path.dirname(__file__), "ldf", "lin_diagnostics.ldf")
+    ldf = parse_ldf(path)
+
+    assert len(ldf.get_diagnostic_frames()) >= 0
+    assert ldf.get_diagnostic_frame('MasterReq').frame_id == LIN_MASTER_REQUEST_FRAME_ID
+    assert ldf.get_diagnostic_frame(LIN_MASTER_REQUEST_FRAME_ID).frame_id == LIN_MASTER_REQUEST_FRAME_ID
+    assert ldf.master_request_frame.frame_id == LIN_MASTER_REQUEST_FRAME_ID
+    assert ldf.slave_response_frame.frame_id == LIN_SLAVE_RESPONSE_FRAME_ID
+
+    assert len(ldf.get_diagnostic_signals()) >= 0
+    assert ldf.get_diagnostic_signal('MasterReqB0').width == 8
+
+    with pytest.raises(LookupError):
+        ldf.get_diagnostic_signal('MasterReqB9')